chore(tel-ui): Move pod state logic to context
